# governance/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RAG Platform API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/external-rags": {
      "get": {
        "tags": [
          "external-rags"
        ],
        "summary": "List External Rags",
        "operationId": "list_external_rags_external_rags_get",
        "parameters": [
          {
            "name": "realm_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Realm Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List External Rags External Rags Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "external-rags"
        ],
        "summary": "Create External Rag",
        "operationId": "create_external_rag_external_rags_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExternalRagCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create External Rag External Rags Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/external-rags/{rag_id}": {
      "patch": {
        "tags": [
          "external-rags"
        ],
        "summary": "Update External Rag",
        "operationId": "update_external_rag_external_rags__rag_id__patch",
        "parameters": [
          {
            "name": "rag_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Rag Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExternalRagUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update External Rag External Rags  Rag Id  Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "external-rags"
        ],
        "summary": "Delete External Rag",
        "operationId": "delete_external_rag_external_rags__rag_id__delete",
        "parameters": [
          {
            "name": "rag_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Rag Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/external-rags/{rag_id}/test": {
      "post": {
        "tags": [
          "external-rags"
        ],
        "summary": "Test External Rag",
        "description": "Send one real probe query so the user can verify the endpoint before\nspending a full experiment run on it. Reports whether it answered at\nall, and whether it returned the optional trace contract.",
        "operationId": "test_external_rag_external_rags__rag_id__test_post",
        "parameters": [
          {
            "name": "rag_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Rag Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Test External Rag External Rags  Rag Id  Test Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExternalRagCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "headers": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Headers",
            "default": {}
          },
          "retrieve_endpoint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retrieve Endpoint"
          },
          "timeout_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout S"
          },
          "request_template": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Template"
          },
          "response_mapping": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Mapping"
          },
          "supported_params": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Supported Params",
            "default": []
          },
          "realm_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Realm Id"
          },
          "default_corpus_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Corpus Id"
          },
          "default_pipeline_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Pipeline Id"
          },
          "default_reranker_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Reranker Id"
          },
          "default_params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Default Params",
            "default": {}
          },
          "uses_realm_resources": {
            "type": "boolean",
            "title": "Uses Realm Resources",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name",
          "url"
        ],
        "title": "ExternalRagCreateRequest"
      },
      "ExternalRagUpdateRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "headers": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headers"
          },
          "retrieve_endpoint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retrieve Endpoint"
          },
          "timeout_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout S"
          },
          "supported_params": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supported Params"
          },
          "default_corpus_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Corpus Id"
          },
          "default_pipeline_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Pipeline Id"
          },
          "default_reranker_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Reranker Id"
          },
          "default_params": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Params"
          },
          "uses_realm_resources": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uses Realm Resources"
          }
        },
        "type": "object",
        "title": "ExternalRagUpdateRequest",
        "description": "All fields optional — PATCH semantics, only provided keys are changed.\nRegistration (POST) has no edit path today; this is the first one, added\nspecifically so default_* can be set/corrected after the fact without\ndeleting and recreating the record (which would also orphan its\nregistered datasets — see services/api_gateway/routers/datasets.py)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}